gprof: Remove useless cast.
Reviewed-on: https://git.eclipse.org/r/32926
Tested-by: Hudson CI

+++ gprof/org.eclipse.linuxtools.gprof/src/org/eclipse/linuxtools/internal/gprof/utils/LEDataInputStream.java
@@ -78,7 +78,7 @@ public class LEDataInputStream extends FilterInputStream implements DataInput {
         (long)(buffer[3]&0xff) << 24 |
         (long)(buffer[2]&0xff) << 16 |
         (long)(buffer[1]&0xff) <<  8 |
-        (long)(buffer[0]&0xff);
+        buffer[0]&0xff;
     }
 
     @Override
